Fix handshake problems I was occasionally seeing issues with exchanging information during the auth state. It turns out that both the bid and public key were often being received before I would read the message. I hadn't really spent much time considering what would happen if we got two messages between reads of the message buffer, and subsequently didn't have a mechanism in place to delineate between the messages.
This change adds code to the UnixSocketNetEngine to add a 2 byte header
and 1 byte footer to all messages. The 2 byte header tells the length of
the message, including the header and footer. The footer is a simple
"0xaa". This isn't perfect. The footer should probably be a checksum,
and there should probably be a means of handling messages that may not
come in one piece. But for now, this fix mitigates the immediate
problem.